+++ src/controllers/UserController.js
@@ -155,8 +155,12 @@ module.exports = {
       }
 
       const user = await User.findByPk(id).catch((err) => {
-        return res.status(400).json({ error: "Registro não encontrado" });
+        return res.status(404).json({ error: "Registro não encontrado" });
       });
+
+      if(user == null ) {
+        return res.status(404).json({ error: "Registro não encontrado" });
+      }
 
       await user.destroy({ where: {id} });
